bug(valPtr): can't dereference a nested pointer
It is currently not possible to dereference a nautilus val that wraps a nested pointer, e.g., `nautilus::val<char**>`. The dereference operator in `val_ptr.hpp` assumes an underlying value that it can return a value to.
It seems to be tricky to add a dereference value that returns a nautilus value of a pointer, since the pointer to the pointer will be a nullptr at trace time. Dereferencing the nullptr leads to a crash.