fix a testcase which relied on old type conversions
prior to the integer promotion change, the function under test would return a val<uint8_t>.
the test was relying on the signed integer value wrapping (which is UB anyway).

The legacy test is fixed by forcing a cast to int8_t, and a new test was added that tests the conversion behavior.